requirements.txt
telethon==1.34.0
py-tgcalls==0.9.7
edge-tts==6.1.10
faster-whisper==0.10.0
openai==1.12.0
httpx==0.27.0
aiohttp==3.9.3
aiofiles==23.2.1
python-dotenv==1.0.1
pydub==0.25.1
numpy==1.26.4
